implement status and address caching. Implement simple design screens

+++ lib/domain/model/weather.dart
@@ -18,6 +18,6 @@ class Weather {
   final int atmosphericPressureHPA;
   final int humidityPercent;
   final double windSpeed;
-  final int windDirection;
+  final double windDirection;
   final WeatherVisualIndicator weatherIcon;
 }
